requirements.txt
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-bigquery==1.28.0
google-cloud-core==1.4.4
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.10
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pytz==2022.7.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.9
